Balance gross trade

Scale gross export and import in all regions to make them equal at the world level.

GROSS_TRADE_ADJUST(
  .DF,
  .MIN_TRADE_PROD_RATIO = 0.01,
  .Reg_VAR = "area_code",
  .GROUP_VAR = c("item_code", "year")
)

Arguments

.DF

An input dataframe with an element col including Import and Export

.MIN_TRADE_PROD_RATIO

Trade will be removed if world total export or import over production is smaller than .MIN_TRADE_PROD_RATIO, 0.01 default value

.Reg_VAR

Region variable name; default is area_code

.GROUP_VAR

Group variable; default is item_code and year

Value

The same dataframe with balanced world export and import.