info

Print and return information about an object.

info(
  object_name,
  gcam_data_map = NULL,
  previous_tracelist = NULL,
  upstream = TRUE,
  downstream = TRUE,
  print = TRUE
)

Arguments

object_name

Name of object to get information about (can be either a data object or a code chunk)

gcam_data_map

A tibble of metadata information; normally a built-in package dataset

previous_tracelist

Information about previous objects printed (if called from dstrace)

upstream

Print and return upstream precursors? Logical

downstream

Print and return downstream dependents? Logical

print

Print to console? Logical

Value

If called directly, returns an entry from GCAM_DATA_MAP; if called from dstrace, a two- element list with linked object and tracelist information. If called and object_name is the name of a code chunk (as opposed to that of a data object), the relevant help page will be pulled up.